move remaining Java files into src; move test deps to main
