+++ unicodetools/src/test/java/org/unicode/propstest/CheckEmojiProps2.java
